+++ CoreLib/PacketTypes.h
@@ -65,7 +65,7 @@ namespace Core {
     //  네트워크 계층에서 패킷을 열어봐야한다면 htonl, htons로 변환해서 써야됨.
     struct PacketHeader {
         uint16_t magic = MAGIC;
-        uint16_t length;
+        uint32_t length;
         uint8_t opcode;
         uint8_t flags = 0x00; // 첫번째 비트는 시뮬레이션 로직인지 나타냄, 0x01 ~ 0x80
     };
